chore(system service): removes unused wireless service
