+++ apps/Cargo.toml
@@ -38,3 +38,4 @@ tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.17", features = ["env-filter"]}
 anyhow = { version = "1.0.75", features = ["backtrace"]}
 tokio = { version = "1.33", features = ["full"] }
+networkmanager = { path = "../commons/networkmanager"}
