+++ apps/settings-app/src/screens/network/saved_network_details.rs
@@ -16,7 +16,7 @@ use mctk_core::{
     Color, Node,
 };
 use mctk_macros::{component, state_component_impl};
-use mechanix_system_dbus_client::wireless::WirelessInfoResponse;
+use networkmanager::network_manager::WirelessInfoResponse;
 
 enum NetworkDetailsMessage {
     openModel(bool),
