+++ services/system/client/dbus/src/proxies/mod.rs
@@ -3,4 +3,3 @@ pub mod display_proxy;
 pub mod hardware_button;
 pub mod host_metrics;
 pub mod security_proxy;
-pub mod wireless_proxy;
