+++ services/system/server/dbus/src/lib.rs
@@ -6,12 +6,4 @@ pub use system_interfaces::{
     bluetooth_event_notification_stream,
     BluetoothBusInterface,
     BluetoothNotificationEvent,
-
-    //wireless interface
-    KnownNetworkListResponse,
-    KnownNetworkResponse,
-    WirelessInfoResponse,
-    WirelessNotificationEvent,
-    WirelessScanListResponse,
-    //power btn
 };
