+++ shell/commons/status-bar/src/modules/wireless/mod.rs
@@ -1,4 +1,2 @@
 pub mod component;
 pub mod errors;
-pub mod handler;
-pub mod service;
